pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "delta-phylo"
version = "0.1.0"
description = "Morphological phylogenetic analyses using DELTA format datasets"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "delta-phylo contributors"}
]
keywords = ["phylogenetics", "DELTA", "morphology", "systematics", "taxonomy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21",
    "pandas>=1.3",
    "biopython>=1.79",
    "matplotlib>=3.4",
    "scipy>=1.7",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "flake8>=6.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.scripts]
delta-phylo = "delta_phylo.cli.main:cli"

[project.urls]
Homepage = "https://github.com/nathanufpb/morfology_analysis"
Repository = "https://github.com/nathanufpb/morfology_analysis"

[tool.setuptools.packages.find]
where = ["."]
include = ["delta_phylo*"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
